Нарезал компоненты
