--- src/pages/main-page.tsx
@@ -1,1 +1,0 @@
-
